Changed MaxLength of admin name from 10 to 20
The main issue I had with DaRT when I first started using it was the fact I could not fit my entire name in the admin name section. I had a friend fix that for me, so it only makes sense that this public version also allows longer names (MaxLength for the admin name has been updated from 10 to 20).